// src/app/models/login-usuario.ts
export class LoginUsuario {

    id?: number;
    username?: string;
    password?: string;
    nombre?: string
    apellido?: string;
    idUsuario?: number;
    idEmpleado?: number;
    roles?: string[] = [];

    // constructor(username: string, password: string) {
    //     this.username = username;
    //     this.password = password;
    // }



}